+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 ### New/Changed Features
+* Download DMD using platform specific archive if available
 * Download DMD from `downloads.dlang.org` (fallback to `ftp.digitalmars.com`)
 * Set a user agent when fetching compilers. The user agent will be in the form
 of `dvm/<version> (<architecture>-<operating system>)`